WIP: implement validateReferer
devide validate-referer.spec.ts
