+++ apps/app/src/server/middlewares/certify-shared-file/validate-referer/index.ts
@@ -1,0 +1,1 @@
+export * from './validate-referer';
